调整题库添加接口

+++ ruoyi-modules/ruoyi-edu/src/main/java/org/dromara/edu/course/domain/CourseCategory.java
@@ -40,7 +40,7 @@ public class CourseCategory extends TenantEntity {
     /**
      * 排序号
      */
-    private Long sortNum;
+    private Integer sortNum;
 
     /**
      * 分类层级
