+++ ruoyi-modules/ruoyi-edu/src/main/java/org/dromara/edu/course/domain/bo/CourseChapterBo.java
@@ -33,7 +33,7 @@ public class CourseChapterBo extends BaseEntity {
     /**
      * 排序号
      */
-    private Long sortNum;
+    private Integer sortNum;
 
     /**
      * 关联课程ID
