+++ ruoyi-modules/ruoyi-edu/src/main/java/org/dromara/edu/course/domain/vo/CourseChapterVo.java
@@ -44,7 +44,7 @@ public class CourseChapterVo implements Serializable {
      * 排序号
      */
     @ExcelProperty(value = "排序号")
-    private Long sortNum;
+    private Integer sortNum;
 
     /**
      * 关联课程ID
