+++ ruoyi-modules/ruoyi-edu/src/main/java/org/dromara/edu/exam/qu/domain/vo/QuestionsVo.java
@@ -44,7 +44,7 @@ public class QuestionsVo implements Serializable {
      * 题目类型：1.单选、2.多选、3.判断、4.简答、5.填空
      */
     @ExcelProperty(value = "题目类型：1.单选、2.多选、3.判断、4.简答、5.填空")
-    private Long questionsType;
+    private Integer questionsType;
 
     /**
      * 正确选项
